move sub-benchmarks into >= go-1.7 only file
