+++ codec_test.go
@@ -35,11 +35,9 @@ package grpc
 
 import (
 	"bytes"
-	"fmt"
 	"sync"
 	"testing"
 
-	"github.com/golang/protobuf/proto"
 	"google.golang.org/grpc/test/codec_perf"
 )
 
@@ -141,81 +139,5 @@ func TestStaggeredMarshalAndUnmarshalUsingSamePool(t *testing.T) {
 		if expectedBody2[i] != v {
 			t.Errorf("expected %v at index %v but got %v", i, expectedBody2[i], v)
 		}
-	}
-}
-
-func setupBenchmarkProtoCodecInputs(b *testing.B) []proto.Message {
-	// small, arbitrary byte slices
-	protoBodies := [][]byte{
-		[]byte("one"),
-		[]byte("two"),
-		[]byte("three"),
-		[]byte("four"),
-		[]byte("five"),
-	}
-	protoStructs := make([]proto.Message, 0)
-
-	for _, p := range protoBodies {
-		ps := &codec_perf.Buffer{}
-		ps.Body = p
-		protoStructs = append(protoStructs, ps)
-	}
-
-	return protoStructs
-}
-
-// The possible use of certain protobuf APIs like the proto.Buffer API potentially involves caching
-// on our side. This can add checks around memory allocations and possible contention.
-// Example run: go test -v -run=^$ -bench=BenchmarkProtoCodec -benchmem
-func BenchmarkProtoCodec(b *testing.B) {
-	i := uint32(0)
-	protoStructs := setupBenchmarkProtoCodecInputs(b)
-	for i < 20 {
-		i += 2
-		func(p int) {
-			b.Run(fmt.Sprintf("BenchmarkProtoCodec_SetParallelism(%v)", p), func(b *testing.B) {
-				codec := &protoCodec{}
-				b.SetParallelism(p)
-				b.RunParallel(func(pb *testing.PB) {
-					benchmarkProtoCodec(codec, protoStructs, pb, b)
-				})
-			})
-		}(1 << i)
-	}
-}
-
-func BenchmarkProtoCodec_Size_1024(b *testing.B) {
-	ps := make([]proto.Message, 1)
-	p := &codec_perf.Buffer{}
-	p.Body = make([]byte, 1024)
-	ps[0] = p
-	benchmarkProtoCodecWithParallelism(b, 65536, ps)
-}
-
-func benchmarkProtoCodecWithParallelism(b *testing.B, p int, protos []proto.Message) {
-	codec := &protoCodec{}
-	b.SetParallelism(p)
-	b.RunParallel(func(pb *testing.PB) {
-		benchmarkProtoCodec(codec, protos, pb, b)
-	})
-}
-
-func benchmarkProtoCodec(codec *protoCodec, protoStructs []proto.Message, pb *testing.PB, b *testing.B) {
-	counter := 0
-	for pb.Next() {
-		counter++
-		ps := protoStructs[counter%len(protoStructs)]
-		fastMarshalAndUnmarshal(codec, ps, b)
-	}
-}
-
-func fastMarshalAndUnmarshal(protoCodec Codec, protoStruct proto.Message, b *testing.B) {
-	marshaledBytes, err := protoCodec.Marshal(protoStruct)
-	if err != nil {
-		b.Errorf("protoCodec.Marshal(_) returned an error")
-	}
-
-	if err := protoCodec.Unmarshal(marshaledBytes, protoStruct); err != nil {
-		b.Errorf("protoCodec.Unmarshal(_) returned an error")
 	}
 }
